[InstagramBridge] Use more reliable endpoint
The bridge failed to provide any feed items for some usernames without
any errors. Use the same endpoint currently used by Instagram's website.

Only use this endpoint if no session data was configured as this change
wasn't tested with it.

The new code also detects when Instagram's rate limit and handles it
appropriately.

In case of errors continue with the same code as before as it's unclear
in which situations this new endpoint might not work.